+++ docs/server/examples/nextjs-app-router/src/app/route.ts
@@ -1,4 +1,4 @@
-import { generateIframeHtml } from "@perspective-software/cross-origin-html-embed/server";
+import { generateIframeHtml } from "@perspective-software/cross-origin-html-embed";
 
 export const dynamic = "force-dynamic"; // defaults to auto
 
@@ -6,7 +6,7 @@ export async function GET() {
   // 🛡️ Here happens the magic
   const iframeHtml = generateIframeHtml({
     allowAllOrigins: false,
-    originWhitelist: ["https://app.myservice.com"],
+    originWhitelist: ["http://localhost:5173"],
     name: "Next.js App Router Example",
   });